deleting functionality finished, added a few params for the new workout form

+++ frontend/src/services/workoutServices.js
@@ -20,19 +20,25 @@ export const createNewWorkout = async (name, description, timeInMinutes, difficu
       };
       try {
         const res = await axios.post('/wor', payload);
+        alert("workout successfully created");
         return res.data;
     } catch (e) {
+        alert("There was a problem creating your workout. please try again.")
         console.error(e);
       }
 }
 
 export const delWorkout = async (nm, wc) => {
-    console.log(nm);
-    console.log(wc);
-    try {
-        await axios.delete('/wor', {params:{name: nm, creatorEmail: wc}});
-        alert("workout deleted");
-    } catch (e) {
-        console.error(e);
-      }
+    const sure = window.confirm("Are you sure you want to delete this workout?");
+    if (sure){
+        try {
+            await axios.delete('/wor', {params:{name: nm, creatorEmail: wc}});
+            alert("Workout deleted. Refresh the page to view changes.");
+        } catch (e) {
+            console.error(e);
+        }
+    }
+    else{
+        alert("Deletion Cancelled");
+    }
 }
